[ufonormalizer] use Unix LF as newline character on all platforms
Since we are writing output file in "text" mode, by default Python uses
the native `os.linesep`, which on Windows is CRLF ("\r\n").

We want all line endings to be normalized as Unix line feeds.
By passing a `newline="\n"` argument to `io.open`, we make sure this is
the case on every platform, including Windows.