ufonormalizer: store modTimes and imageReferences in layerInfo's "lib", not in top-level dict
http://www.unifiedfontobject.org/versions/ufo3/glyphs.html#layerinfo

According to the UFO3 spec, the layerinfo.plist data is a dictionary containing "color" definition and a "lib" sub-dictionary. The ufonormalizer was confusing the top-level dictionary with the lib dictionary, and overwriting the former with the content of the latter.
Instead, modTimes and imageReferences must be read/written from/in the inner "lib" dict of layerinfo.plist.

Later on, we should write tests for this, as well as the whole `normalizeGlyphsDirectory` function.