fix running `coverage src/ufonormalizer.py -t`
I moved the tests to a separate folder, so I need to
extened the sys.path to include the tests folder for
unittest.main to be able to import test_ufonormalizer.py.

I also changed the folder which is "globbed" for *.ufo.
This is now 'tests/data/' (updated .gitignore accordingly).